个人预约修改信息
个人预约修改信息

+++ HealthyCertificate/ProjectClass/Model/WaitToCheckModel/CustomerTest.h
@@ -92,11 +92,11 @@
 /**
  * 预约经度
  */
-@property (nonatomic, assign) double regPosLO;
+@property (nonatomic, assign) long long regPosLO;
 /**
  * 预约纬度
  */
-@property (nonatomic, assign) double regPosLA;
+@property (nonatomic, assign) long long regPosLA;
 /**
  * 预约地址
  */
